+++ Test/Unittest/blackbox/flask/test_static_resources.py
@@ -1,4 +1,5 @@
 from json import loads
+from pathlib import Path
 from typing import Any
 
 from flask_unittest import ClientTestCase
@@ -10,6 +11,7 @@ _APP_CONFIG = {
     "repositories_config": {
         "repository_type": "memory",
     },
+    "resources_folder_path": Path(__file__).parents[4].joinpath("Source/Resources/").resolve().as_posix(),
 }