+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Changed the `reality` component of Locations' and Events' `span` attribute to be a set of integers instead of a range
   - Changed served Json representation to be a sorted integer array
   - Changed API Spec to align with integer array changes
+- Changed metadata attribute to reject empty strings for either keys or values
 
 ---